requirements.txt
Flask==3.0.2
opencv-python-headless==4.8.1.78
pillow==10.0.1
imagehash==4.3.1
numpy==1.26.4
pytubefix==7.1.0
python-pptx==1.0.2
requests==2.31.0